server/requirements.txt
Flask>=3.0
flask-cors>=4.0
pymongo>=4.9
bcrypt>=4.1
PyJWT>=2.9
python-dotenv>=1.0
tensorflow>=2.15
Pillow>=10.0
numpy>=1.26
deepface>=0.0.95
opencv-python>=4.10.0.84 